Refactor HTTP client and API handler***
***Adjust concurrency management based on metrics
